pyproject.toml
[project]
name = "vqms"
version = "1.0.0"
description = "VQMS — Vendor Query Management System Agentic AI Platform"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pydantic-settings",
    # AI / LLM — Amazon Bedrock + LangChain/LangGraph
    "boto3",
    "botocore",
    "langchain>=0.3",
    "langchain-aws",
    "langchain-community",
    "langgraph>=0.2",
    "langsmith",
    # Database — PostgreSQL + pgvector
    "asyncpg",
    "psycopg2-binary",
    "pgvector",
    "sqlalchemy[asyncio]",
    "alembic",
    # Cache — Redis
    "redis[hiredis]",
    # API & Web — External Service Adapters
    "httpx",
    "aiohttp",
    "requests",
    "msal",
    "simple-salesforce",
    "pysnow",
    # Email Parsing
    "python-multipart",
    "email-validator",
    # Observability & Logging
    "structlog",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-otlp",
    "aws-xray-sdk",
    # Templating & Prompts
    "jinja2",
    # Security & Compliance
    "cryptography",
    "python-jose[cryptography]",
    # Utilities
    "python-dotenv",
    "pyyaml",
    "tenacity",
    "python-dateutil",
    "orjson",
    "beautifulsoup4>=4.14.3",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "black",
    "isort",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "moto[s3,sqs,events,stepfunctions]",
    "ragas",
    "deepeval",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
